requirements.txt
flask==2.3.2
gunicorn==20.1.0
opencv-python-headless==4.7.0.72
numpy==1.24.3
tensorflow-cpu==2.12.0